+++ src/vscode-mock.ts
@@ -7,7 +7,7 @@ const mockConfig = {
     get(key: string, defaultValue?: any): any {
         const defaults:any = {
             'formatRightMargin': 70,
-            'formatCase': '*UPPER',
+            'convertCmdAndParmNameCase': '*UPPER',
             'formatLabelPosition': 2,
             'formatCmdPosition': 14,
             'formatKwdPosition': 25,